Move open_connection tests to correct module
